runner: split into runner/ subpackage (executor + progress)
